+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Pre-release
 
 - Remove base64 encoding approach to saving images to filesystem (wasn't properly functional)
+- Remove requirement to set `IMAGE_STORAGE_DIRECTORY` environment variable in favor of reasonable defaults per OS
 
 ## 0.0.3